@@ -16,8 +16,8 @@
     "lodash.debounce": "^4.0.8",
     "normalizr": "^3.4.0",
     "ramda": "0.26.1",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-helmet": "^5.2.1",
     "react-pluralize": "^1.6.3",
     "react-redux": "^7.1.0",
@@ -35,9 +35,9 @@
     "eslint-config-prettier": "^6.0.0",
     "eslint-config-standard": "^13.0.1",
     "eslint-config-standard-react": "^8.0.0",
-    "eslint-plugin-flowtype": "^3.13.0",
+    "eslint-plugin-flowtype": "^4.2.0",
     "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jest": "^22.14.1",
+    "eslint-plugin-jest": "^22.15.0",
     "eslint-plugin-node": "^9.1.0",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-ramda": "^2.0.0",
@@ -51,8 +51,8 @@
     "path": "^0.12.7",
     "pre-commit": "^1.2.2",
     "prettier": "^1.18.2",
-    "react-scripts": "3.0.1",
-    "react-test-renderer": "^16.8.6",
+    "react-scripts": "3.1.0",
+    "react-test-renderer": "^16.9.0",
     "redux-mock-store": "^1.5.3",
     "rosie": "^2.0.1",
     "source-map-explorer": "^2.0.1",
